feat(SPV-848): package notification moved to engine
