+++ engine/models_internal.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"time"
 
-	"github.com/bitcoin-sv/spv-wallet/engine/notifications"
 	"github.com/bitcoin-sv/spv-wallet/engine/spverrors"
 )
 
@@ -159,23 +158,23 @@ func incrementField(ctx context.Context, model ModelInterface, fieldName string,
 }
 
 // notify about an event on the model
-func notify(eventType notifications.EventType, model interface{}) {
-	// run the notifications in a separate goroutine since there could be significant network delay
-	// communicating with a notification provider
+// func notify(eventType notifications.EventType, model interface{}) {
+// 	// run the notifications in a separate goroutine since there could be significant network delay
+// 	// communicating with a notification provider
 
-	go func() {
-		m := model.(ModelInterface)
-		if client := m.Client(); client != nil {
-			if n := client.Notifications(); n != nil {
-				if err := n.Notify(
-					context.Background(), m.GetModelName(), eventType, model, m.GetID(),
-				); err != nil {
-					client.Logger().Error().Msgf("failed notifying about %s on %s: %s", string(eventType), m.GetID(), err.Error())
-				}
-			}
-		}
-	}()
-}
+// 	go func() {
+// 		m := model.(ModelInterface)
+// 		if client := m.Client(); client != nil {
+// 			if n := client.Notifications(); n != nil {
+// 				if err := n.Notify(
+// 					context.Background(), m.GetModelName(), eventType, model, m.GetID(),
+// 				); err != nil {
+// 					client.Logger().Error().Msgf("failed notifying about %s on %s: %s", string(eventType), m.GetID(), err.Error())
+// 				}
+// 			}
+// 		}
+// 	}()
+// }
 
 /*
 // setFieldValueByJSONTag will parse the struct looking for the field (json tag) and updating the value if found
