+++ engine/notifications/notifications.go
@@ -1,60 +1,70 @@
-// Package notifications is a basic internal notifications module
 package notifications
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"net/http"
+	"sync"
 )
 
-// GetWebhookEndpoint will get the configured webhook endpoint
-func (c *Client) GetWebhookEndpoint() string {
-	return c.options.config.webhookEndpoint
+// Event - event type
+type Event any
+
+const lengthOfInputChannel = 100
+
+// Notifications - service for sending events to multiple notifiers
+type Notifications struct {
+	inputChannel   chan Event
+	outputChannels *sync.Map //[string, chan Event]
 }
 
-// Notify will create a new notification event
-func (c *Client) Notify(ctx context.Context, modelType string, eventType EventType,
-	model interface{}, id string) error {
+// AddNotifier - add notifier by key
+func (n *Notifications) AddNotifier(key string, ch chan Event) {
+	n.outputChannels.Store(key, ch)
+}
 
-	if len(c.options.config.webhookEndpoint) == 0 {
-		if c.IsDebug() {
-			c.Logger().Info().Msgf("NOTIFY %s: %s - %v", eventType, id, model)
-		}
-	} else {
-		jsonData, err := json.Marshal(map[string]interface{}{
-			"event_type": eventType,
-			"id":         id,
-			"model":      model,
-			"model_type": modelType,
-		})
-		if err != nil {
-			return err
-		}
+// RemoveNotifier - remove notifier by key
+func (n *Notifications) RemoveNotifier(key string) {
+	n.outputChannels.Delete(key)
+}
 
-		var req *http.Request
-		if req, err = http.NewRequestWithContext(ctx,
-			http.MethodPost,
-			c.options.config.webhookEndpoint,
-			bytes.NewBuffer(jsonData),
-		); err != nil {
-			return err
-		}
+// Notify - send event to all notifiers
+func (n *Notifications) Notify(event Event) {
+	n.inputChannel <- event
+}
 
-		var response *http.Response
-		if response, err = c.options.httpClient.Do(req); err != nil {
-			return err
-		}
-		defer func() {
-			_ = response.Body.Close()
-		}()
-
-		if response.StatusCode != http.StatusOK {
-			// todo queue notification for another try ...
-			c.Logger().Error().Msgf("received invalid response from notification endpoint: %d",
-				response.StatusCode)
+// exchange - exchange events between input and output channels, uses fan-out pattern
+func (n *Notifications) exchange(ctx context.Context) {
+	for {
+		select {
+		case event := <-n.inputChannel:
+			n.outputChannels.Range(func(_, value any) bool {
+				ch := value.(chan Event)
+				n.sendEventToChannel(ch, event)
+				return true
+			})
+		case <-ctx.Done():
+			return
 		}
 	}
+}
 
-	return nil
+// sendEventToChannel - non blocking send event to channel
+func (n *Notifications) sendEventToChannel(ch chan Event, event Event) {
+	select {
+	case ch <- event:
+		// Successfully sent event
+	default:
+		// Channel is full, skip sending event
+	}
+}
+
+// NewNotifications - creates a new instance of Notifications
+func NewNotifications(ctx context.Context) *Notifications {
+	n := &Notifications{
+		inputChannel:   make(chan Event, lengthOfInputChannel),
+		outputChannels: new(sync.Map),
+	}
+
+	go n.exchange(ctx)
+
+	return n
 }
