+++ engine/sync_tx_service.go
@@ -12,7 +12,6 @@ import (
 	"github.com/bitcoin-sv/go-paymail"
 	"github.com/bitcoin-sv/spv-wallet/engine/chainstate"
 	"github.com/bitcoin-sv/spv-wallet/engine/datastore"
-	"github.com/bitcoin-sv/spv-wallet/engine/notifications"
 	"github.com/bitcoin-sv/spv-wallet/engine/spverrors"
 )
 
@@ -156,7 +155,7 @@ func broadcastSyncTransaction(ctx context.Context, syncTx *SyncTransaction) erro
 	}
 
 	// Fire a notification
-	notify(notifications.EventTypeBroadcast, syncTx)
+	// notify(notifications.EventTypeBroadcast, syncTx)
 
 	return nil
 }
